% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis-leaflet.R
\name{vis_leaflet}
\alias{vis_leaflet}
\title{Visualize with leaflet}
\usage{
vis_leaflet(variable, indicator, year_slice, palette = c("RdYlGn"),
  scale = c("numeric", "quantile"), show_real_map = TRUE,
  show_labels = TRUE, legend_title = TRUE, reverse = FALSE)
}
\description{
Visualize with leaflet
}
